+++ apps/docs/src/app/core/component-docs/product-switch/product-switch-docs.component.ts
@@ -6,8 +6,8 @@ import * as smallProductSwitchTs from '!raw-loader!./examples/product-switch-sma
 import * as productSwitchDndH from '!raw-loader!./examples/product-switch-dnd-example.component.html';
 import * as productSwitchDndTs from '!raw-loader!./examples/product-switch-dnd-example.component.ts';
 
-import * as listProductSwitchH from '!raw-loader!./examples/product-switch-list/product-switch-list.component.html';
-import * as listProductSwitchTs from '!raw-loader!./examples/product-switch-list/product-switch-list.component.ts';
+import * as listProductSwitchH from '!raw-loader!./examples/product-switch-list/product-switch-list-example.component.html';
+import * as listProductSwitchTs from '!raw-loader!./examples/product-switch-list/product-switch-list-example.component.ts';
 import { ExampleFile } from '../../../documentation/core-helpers/code-example/example-file';
 
 @Component({
@@ -18,22 +18,30 @@ export class ProductSwitchDocsComponent {
     productSwitchBasicExample: ExampleFile[] = [
         {
             language: 'html',
-            code: productSwitchDndH
+            code: productSwitchDndH,
+            fileName: 'product-switch-dnd-example'
+
         },
         {
             language: 'typescript',
-            code: productSwitchDndTs
+            code: productSwitchDndTs,
+            fileName: 'product-switch-dnd-example',
+            component: 'ProductSwitchDndExampleComponent'
         }
     ];
 
     productSwitchSmallExample: ExampleFile[] = [
         {
             language: 'html',
-            code: smallProductSwitchH
+            code: smallProductSwitchH,
+            fileName: 'product-switch-small-example',
+
         },
         {
             language: 'typescript',
-            code: smallProductSwitchTs
+            code: smallProductSwitchTs,
+            fileName: 'product-switch-small-example',
+            component: 'ProductSwitchSmallExampleComponent'
         }
     ];
 
@@ -41,10 +49,13 @@ export class ProductSwitchDocsComponent {
         {
             language: 'html',
             code: listProductSwitchH,
+            fileName: 'product-switch-list-example',
         },
         {
             language: 'typescript',
-            code: listProductSwitchTs
+            code: listProductSwitchTs,
+            fileName: 'product-switch-list-example',
+            component: 'ProductSwitchListComponent'
         }
     ];