+++ apps/docs/src/app/documentation/core-helpers/docs-section-title/docs-section-title.component.ts
@@ -1,13 +1,15 @@
-import { Component, OnInit, Input, ElementRef, ViewChild, AfterViewInit, TemplateRef } from '@angular/core';
-import { Renderer } from 'marked';
-import { ActivatedRoute } from '@angular/router';
+import { Component, OnInit, Input, ElementRef, ViewChild, AfterViewInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
     selector: 'fd-docs-section-title',
     template: `
         <h2 [id]="id" #title class="docs-header-link">
-            <a class="docs-markdown-a" [attr.aria-describedby]="id" href="/fundamental-ngx/{{ componentName }}#{{ id }}">
-                <span class="sap-icon--chain-link"></span>
+            <a class="docs-markdown-a"
+               [attr.aria-describedby]="id"
+               [routerLink]="'/' + currentLibrary + '/' + componentName"
+               [fragment]="id">
+                <fd-icon [glyph]="'chain-link'"></fd-icon>
             </a>
             <ng-content></ng-content>
         </h2>
@@ -15,25 +17,50 @@ import { ActivatedRoute } from '@angular/router';
     styleUrls: ['./docs-section-title.component.scss']
 })
 export class DocsSectionTitleComponent implements OnInit, AfterViewInit {
+
+    @ViewChild('title', { read: ElementRef, static: false })
+    sectionTitle: ElementRef;
+
+    @Input()
+    id: string = '';
+
+    @Input()
+    componentName: string = '';
+
+    readonly currentLibrary: 'core' | 'platform' = null;
+
     private idFromUrl: any;
 
-    @ViewChild('title', { read: ElementRef, static: false  }) sectionTitle: ElementRef;
+    constructor(
+        private activatedRoute: ActivatedRoute,
+        private router: Router
+    ) {
+        this.currentLibrary = this.router.url && this.router.url.includes('core/') ? 'core' : 'platform';
+    }
 
-    @Input() id: string = '';
-
-    @Input() componentName: string = '';
-
-    constructor(private activatedRoute: ActivatedRoute) {}
-
-    ngOnInit() {
+    ngOnInit(): void {
         this.activatedRoute.fragment.subscribe(fragment => {
             this.idFromUrl = fragment;
+            this.handleUrlFragment();
         });
     }
 
     ngAfterViewInit(): void {
-        if (this.sectionTitle.nativeElement.firstChild.hash === '#' + this.idFromUrl) {
-            this.sectionTitle.nativeElement.scrollIntoView();
+        this.handleUrlFragment();
+    }
+
+    private handleUrlFragment(): void {
+        if (this.sectionTitle) {
+            if (this.id === this.idFromUrl) {
+                this.sectionTitle.nativeElement.scrollIntoView(true);
+                this.addOffset();
+            }
+        }
+    }
+
+    private addOffset(): void {
+        if (document.getElementById('page-content')) {
+            document.getElementById('page-content').scrollTop -= 30;
         }
     }
 }