+++ libs/platform/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/platform.module';
 export * from './lib/components/action-bar/public_api';
+export * from './lib/components/menu/public_api';