+++ CHANGELOG.md
@@ -2,6 +2,71 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.13.0-rc.0"></a>
+# [0.13.0-rc.0](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.9...v0.13.0-rc.0) (2019-10-31)
+
+
+### Features
+
+* **platform:** Menu as a standalone component ([#1390](https://github.com/SAP/fundamental-ngx/issues/1390)) ([99ea344](https://github.com/SAP/fundamental-ngx/commit/99ea344)), closes [#1264](https://github.com/SAP/fundamental-ngx/issues/1264) [#1264](https://github.com/SAP/fundamental-ngx/issues/1264) [#1264](https://github.com/SAP/fundamental-ngx/issues/1264) [#1264](https://github.com/SAP/fundamental-ngx/issues/1264) [#1264](https://github.com/SAP/fundamental-ngx/issues/1264) [#1264](https://github.com/SAP/fundamental-ngx/issues/1264)
+
+
+
+<a name="0.12.1-rc.9"></a>
+## [0.12.1-rc.9](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.8...v0.12.1-rc.9) (2019-10-29)
+
+
+
+<a name="0.12.1-rc.8"></a>
+## [0.12.1-rc.8](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.7...v0.12.1-rc.8) (2019-10-28)
+
+
+
+<a name="0.12.1-rc.7"></a>
+## [0.12.1-rc.7](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.6...v0.12.1-rc.7) (2019-10-28)
+
+
+### Bug Fixes
+
+* stackblitz for product-switch ([#1486](https://github.com/SAP/fundamental-ngx/issues/1486)) ([24606c9](https://github.com/SAP/fundamental-ngx/commit/24606c9))
+
+
+
+<a name="0.12.1-rc.6"></a>
+## [0.12.1-rc.6](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.5...v0.12.1-rc.6) (2019-10-28)
+
+
+### Bug Fixes
+
+* add DnD styling improvements to Product Switch ([#1490](https://github.com/SAP/fundamental-ngx/issues/1490)) ([7ed1839](https://github.com/SAP/fundamental-ngx/commit/7ed1839))
+
+
+
+<a name="0.12.1-rc.5"></a>
+## [0.12.1-rc.5](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.4...v0.12.1-rc.5) (2019-10-28)
+
+
+
+<a name="0.12.1-rc.4"></a>
+## [0.12.1-rc.4](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.3...v0.12.1-rc.4) (2019-10-28)
+
+
+### Bug Fixes
+
+* Modal height example issues. Close icon fix. ([#1483](https://github.com/SAP/fundamental-ngx/issues/1483)) ([3a0d185](https://github.com/SAP/fundamental-ngx/commit/3a0d185))
+
+
+
+<a name="0.12.1-rc.3"></a>
+## [0.12.1-rc.3](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.2...v0.12.1-rc.3) (2019-10-28)
+
+
+
+<a name="0.12.1-rc.2"></a>
+## [0.12.1-rc.2](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.1...v0.12.1-rc.2) (2019-10-25)
+
+
+
 <a name="0.12.1-rc.1"></a>
 ## [0.12.1-rc.1](https://github.com/SAP/fundamental-ngx/compare/v0.12.1-rc.0...v0.12.1-rc.1) (2019-10-25)